+++ cloudsim-plus-testbeds/src/main/java/hostFaultInjection/HostFaultInjectionRunner.java
@@ -38,6 +38,9 @@ class HostFaultInjectionRunner extends ExperimentRunner<HostFaultInjectionExperi
      */
     private List<Double> percentageOfBrokersMeetingAvailability;
 
+    /**
+     * Average number of VMs for each existing Host.
+     */
     private List<Double> ratioVmsPerHost;
 
 
@@ -54,7 +57,7 @@ class HostFaultInjectionRunner extends ExperimentRunner<HostFaultInjectionExperi
      */
     public static void main(String[] args) {
         new HostFaultInjectionRunner(true, 1475098589732L)
-            .setSimulationRuns(300)
+            .setSimulationRuns(10)
             .setNumberOfBatches(5) //Comment this or set to 0 to disable the "Batch Means Method"
             .setVerbose(true)
             .run();
@@ -84,7 +87,7 @@ class HostFaultInjectionRunner extends ExperimentRunner<HostFaultInjectionExperi
      */
     private void afterExperimentFinish(HostFaultInjectionExperiment exp) {
         availability.add(exp.getFaultInjection().availability() * 100);
-        percentageOfBrokersMeetingAvailability.add(exp.getPercentageOfAvailabilityThatMeetingTheSla());
+        percentageOfBrokersMeetingAvailability.add(exp.getPercentageOfAvailabilityThatMeetingTheSla() * 100);
         ratioVmsPerHost.add(exp.getRatioVmsPerHost());
     }
