+++ cloudsim-plus-testbeds/src/main/java/org/cloudsimplus/sla/TestWithoutAlgorithmExample.java
@@ -70,6 +70,7 @@ import org.cloudsimplus.autoscaling.HorizontalVmScalingSimple;
 import org.cloudsimplus.builders.tables.CloudletsTableBuilder;
 import org.cloudsimplus.listeners.EventInfo;
 import org.cloudsimplus.listeners.EventListener;
+import org.cloudsimplus.sla.readJsonFile.slaMetricsJsonFile.CpuUtilization;
 import org.cloudsimplus.sla.readJsonFile.slaMetricsJsonFile.TaskTimeCompletion;
 import org.cloudsimplus.sla.readJsonFile.slaMetricsJsonFile.SlaReader;
 import org.cloudsimplus.sla.readJsonFile.slaMetricsJsonFile.Availability;
@@ -141,11 +142,11 @@ public class TestWithoutAlgorithmExample {
         rt.checkTaskTimeCompletionSlaContract();
         taskTimeCompletionSlaContract = rt.getMaxValueTaskTimeCompletion();
 
-        Availability.CpuUtilization cpu = new Availability.CpuUtilization(slaReader);
+        CpuUtilization cpu = new CpuUtilization(slaReader);
         cpu.checkCpuUtilizationSlaContract();
         cpuUtilizationSlaContract = cpu.getMaxValueCpuUtilization();
 
-//        simulation.addOnClockTickListener(this::createNewCloudlets);
+//       simulation.addOnClockTickListener(this::createNewCloudlets);
         simulation.addOnClockTickListener(this::printVmsCpuUsage);
 
         createDatacenter();
