+++ cloudsim-plus-testbeds/src/main/java/org/cloudsimplus/sla/tasktimecompletion/CloudletTaskTimeCompletionMinimizationExperiment.java
@@ -62,6 +62,7 @@ import org.cloudbus.cloudsim.vms.Vm;
 import org.cloudbus.cloudsim.vms.VmSimple;
 import org.cloudsimplus.builders.tables.CloudletsTableBuilder;
 import org.cloudsimplus.listeners.EventInfo;
+import org.cloudsimplus.sla.readJsonFile.slaMetricsJsonFile.CpuUtilization;
 import org.cloudsimplus.sla.readJsonFile.slaMetricsJsonFile.TaskTimeCompletion;
 import org.cloudsimplus.sla.readJsonFile.slaMetricsJsonFile.SlaReader;
 import static org.cloudsimplus.sla.tasktimecompletion.CloudletTaskTimeCompletionMinimizationRunner.CLOUDLETS;
@@ -128,7 +129,7 @@ public final class CloudletTaskTimeCompletionMinimizationExperiment extends Simu
             rt.checkTaskTimeCompletionSlaContract();
             taskTimeCompletionSlaContract = rt.getMaxValueTaskTimeCompletion();
 
-            Availability.CpuUtilization cpu = new Availability.CpuUtilization(slaReader);
+            CpuUtilization cpu = new CpuUtilization(slaReader);
             cpu.checkCpuUtilizationSlaContract();
             cpuUtilizationSlaContract = cpu.getMaxValueCpuUtilization();
