+++ cloudsim-plus-testbeds/src/main/resources/m4.large
@@ -1,0 +1,6 @@
+{
+  "instanceName": "m4.large",
+  "vCPU":2,
+  "memoryInMB":8192,
+  "pricePerHour":0.192
+}
