+++ cloudsim-plus-testbeds/src/main/resources/p2.xlarge
@@ -1,0 +1,6 @@
+{
+  "instanceName": "p2.xlarge",
+  "vCPU":4,
+  "memoryInMB":62464,
+  "pricePerHour":1.084
+}
