+++ cloudsim-plus-testbeds/src/main/resources/t2.large
@@ -1,0 +1,6 @@
+{
+  "instanceName": "t2.large",
+  "vCPU":2,
+  "memoryInMB":8192,
+  "pricePerHour":0.122
+}
