+++ cloudsim-plus-testbeds/src/main/resources/t2.nano
@@ -1,0 +1,7 @@
+{
+  "instanceName": "t2.nano",
+  "vCPU": 1,
+  "memoryInMB": 512,
+  "pricePerHour": 0.0082
+}
+
